Refactor attachment and message effects logic
Removed AttachmentFinalizer and MessageEffects services, moving their logic into S3Service and RedisService respectively. Updated interfaces and usages throughout the WebSocket message flow to use the new methods on S3Service and RedisService. This simplifies the service layer and centralizes related logic.
