+++ server/Express/domain/interfaces/storageInterface/IS3Service.ts
@@ -1,3 +1,5 @@
+import { FinalizeResult } from "../../services/storageServices/S3Service";
+
 export interface IS3Service {
   uploadBuffer(
     buffer: Buffer,
@@ -7,4 +9,5 @@ export interface IS3Service {
   copyObject(srcKey: string, dstKey: string): Promise<void>;
   deleteObject(key: string): Promise<void>;
   publicUrl(key: string): string;
+  finalize(roomId: string, userId: string, ts: number, attachments: string[]): Promise<FinalizeResult>;
 }
